Support tagging ChatGPT bot via its primary email
Right now to ask questions to ChatGPT it's required to provide the bot
username in the comment.

In the regular workflow with Gerrit, typing `@` following by a single
letter opens  user auto-suggestion dropdown. From which the user can
be selected. This also applies to the ChatGPT user.

Unfortunately, this won't trigger the plugin action,
as auto-suggestion is inserting the primary email address of the user.

This change updates the regular expression for finding questions to
ChatGPT to work with also with the ChatGPT user's prefferred email
address.

To make it work, we extend the `Configuration` class to also include
ChatGPT user email address. This is because that the
`ConfigurationCreator` class is the last place where we can use Guice
injections. With that in place, we can then update the bot mention
pattern in `ClientMessage`.

Change-Id: I89500741600bc584ee749f3ff0cea2f7cc23f292