Tighten email address constraint for tagging
With the preferred email address of ChatGPT users now available to be
used for directing queries to ChatGPT, a strict matching constraint for
these email addresses is implemented.
Specifically, the format "@<ChatGPT_username>@<any_domain>" will no
longer trigger the ChatGPT's replies.
This replaces the previous regular expression that matched any email
starting with the username (typically "gpt") assuming it was the usual
format for ChatGPT email addresses.

Change-Id: I6228c83f4773d607cb936fa51ca1d0d4d6def66e
Signed-off-by: Patrizio <patrizio.gelosi@amarulasolutions.com>