User direct GerritApi calls in GerritClientAccount
Modify both `getAccountId` and `getAccountGroups` to use direct
`GerritApi` calls. Note that it still extends the `GerritClientBase` as
it is a parent for several implementations (that will be refactored
later).

The `ChatGptReviewTestBase` was modified to mock direct `GerritApi`
calls related to accounts. As a result all tests are still passing.

Change-Id: Iaee041a4c38ae842ba4707841c782dd1fb33f664