Migrate GerritClientReview to direct GerritApi
The `GerritClientReview` uses ReviewInput to post ChatGPT review on a
change in question.
Note that in order to maintain existing unit tests the expected
body content is read and converted back to compact JSON representation
so that it is comparable with value that is captured from mock (that is
also turned to compact JSON) as `ReviewInput` has no `equals` method
implementation that could be used otherwise.

Bug: Isssue 336577745
Change-Id: I41bd1ef4d1f978843b29bfbd4942dda5d96d135b