Fix detection of the last posted comments for ChatGPT
When trying to detect the last comments posted to a
change, consider the ones created in the past 60s so to
allow a bit of latency between the creation of the change notes
and the notification to the stream events listener.

This is a temporary solution which is keeping the current
approach. A much better fix would be refactoring the way
the comments are retrieved and fetching them directly
from NoteDb, because the comment author is also stored
in the author id of the /meta commit.

Change-Id: I9147741ecaf2c57cb30c1a4c0950c688b720d223