Enforce thread history lookup in stateful prompts
A new rule has been implemented in both the review and request stateful
prompts, explicitly requiring ChatGPT to consider the history of
messages exchanged in the thread. Additionally, a list of mandatory
rules has been added to the review prompt to ensure adherence to each
requirement. Among these, the rule stipulating that responses must be
provided as a JSON object is reiterated.

Change-Id: I2ed9b0372e7494fa967f7876f274f4b841c42c29
Signed-off-by: Patrizio <patrizio.gelosi@amarulasolutions.com>