Support selective log level override configuration
Added a new configuration setting, `selectiveLogLevelOverride`, which
enables the logging of specific messages below the current log level
threshold.
This feature allows selected messages with lower log levels to be
included in the log without changing the overall log level to DEBUG,
preventing an overload of DEBUG messages from sources such as gerrit
and other plugins.
Messages to be elevated can be selected based on their class and/or
content.

Change-Id: Ibd202bb9f4a923d5bf2d98514e2e44d2c0e84943
Signed-off-by: Patrizio <patrizio.gelosi@amarulasolutions.com>