Introduce `forget_thread` command
Added a new command to delete the thread history in stateful mode. A
thread documents all previous interactions and reviews by ChatGPT in a
given Change Set, and its history can be erased using this command. This
feature is useful to ensure that ChatGPT does not simply repeat previous
responses in stateful mode, especially after changes to configuration
parameters.

Change-Id: Ib8e52916b260c3c7fb87e4213ca1bd9786ab8654
Signed-off-by: Patrizio <patrizio.gelosi@amarulasolutions.com>