Support command syntax checking
With this commit, any message to ChatGPT that contains a string starting
with a slash followed by an alphanumeric character will be interpreted
as a command. If the command is invalid, a system message will be
displayed.
A command is considered invalid if it is not recognized by the ChatGPT
plugin or if any of its options are unsupported.

Change-Id: I83641cc9bc20767b7e5a951883bef3855a384502
Signed-off-by: Patrizio <patrizio.gelosi@amarulasolutions.com>