Enhance selective log level override management
The configuration for selective log level override is now handled as a
list of strings, offering several benefits:
- Simpler configuration, allowing each filter to be entered on a
separate line in `.config` files.
- Dynamic selective log filters are appended to project-scoped filters
instead of being overridden.
- Improved visualization, as each filter is displayed on a separate line
in the configuration panels in the UI.
- Increased maintainability, with custom management replaced by the
standard List setting management also used for directive settings.

NOTE: The hash symbol ("#") has been replaced by the pipe ("|") as the
separator between the class name filter and the log message filter in
the filter syntax.

Change-Id: I7a561589ee4daaad6c8025f6fcfb74f116af7408
Signed-off-by: Patrizio <patrizio.gelosi@amarulasolutions.com>