Unify `/dump_*` commands into `/show`
The `/dump_config` command is now executed through the `--config`
option, and the `/dump_stored_data` command is executed through the
`--local_data` option of the new `/show` command.
This update simplifies the user experience and prepares for the addition
of other objects to be displayed without the need for dedicated
commands.

Change-Id: I41a28e62749857e8f1568676bdcb4029b355ff5a
Signed-off-by: Patrizio <patrizio.gelosi@amarulasolutions.com>